+++ Assets/Scripts/TrajectoryPlanner/ProbeManager.cs
@@ -1,6 +1,6 @@
 using System;
 using System.Collections.Generic;
-using SensapexLink;
+using EphysLink;
 using TMPro;
 using TrajectoryPlanner;
 using UnityEngine;
