+++ Assets/Tests/Tests.asmdef
@@ -4,7 +4,7 @@
     "references": [
         "UnityEngine.TestRunner",
         "UnityEditor.TestRunner",
-        "trajectoryplanner.sensapexlink"
+        "trajectoryplanner.ephyslink"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
